update UI (aka "draw new circles") even if video playback is paused
This means we cannot always fetch a new frame at each loop iteration but have to respect whether we are in a simple "just update the UI" loop (in pause mode) or regularly fetch new frames.
Note that the UI update rate still depends on the FPS if not in pause mode (as before). That means, with low FPS you also have a lag between clicking and drawing circles.