load and display old annotations for the video when reopening (read-only)
This makes it easier to annotate more than one dance in a video. Note that the loaded annotations are read-only. However, it would now be trivial to also allow editing them.
And yes, when saving lists of values in columns of a CSV file as strings, you get parsing an abstract syntax tree for loading them again...